Extract configuration and generation to separate modules. NFC
